bter faster

+++ src/aoc-init.ts
@@ -1,52 +1,65 @@
 import * as readline from 'readline';
 import * as settings from './settings';
+const path = require('path');
 const fs = require('fs');
 const https = require('https');
 
 const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout
-  });
+  input: process.stdin,
+  output: process.stdout
+});
 
 const _doInit = (answer) => {
   const options = {
-     host: `adventofcode.com`,
-     port:443,
-     method: 'GET',
-     path:`/${settings.Settings.YEAR}/day/${answer}/input`, // 9-11-2022 19:29:58
-     headers: { 'cookie': process.env["AOC_COOKIE"] }
+    host: `adventofcode.com`,
+    port: 443,
+    method: 'GET',
+    path: `/${settings.Settings.YEAR}/day/${answer}/input`, // 9-11-2022 19:29:58
+    headers: {
+      cookie: process.env["AOC_COOKIE"]
+      , 'user-agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/107.0.0.0 Safari/537.36'
+    }
   };
 
+  // general variables
   const zeroPadded = `0${answer}`.substring(answer.length - 1);
   const name = `day${zeroPadded}`;
+
+  // typescript source code file
+  let localCodePath = `${__dirname}\\..\\src\\${settings.Settings.YEAR}\\${name}`;
+  if (!fs.existsSync(localCodePath)) {
+    fs.mkdirSync(localCodePath, { "recursive": true });
+  }
+  localCodePath += `\\code.ts`;
+  if (fs.existsSync(localCodePath)) {
+    console.warn(`[WARN] Code file ${path.resolve(localCodePath)} already exists! Please remove manually.`);
+  } else {
+    let code = fs.readFileSync(`${__dirname}\\..\\src\\code-template.ts`).toString();
+    code = code.replace(/day00/g, name);
+    fs.writeFile(localCodePath, code, (err) => {
+      if (err)
+        console.error(`[ERR] Error: ${err}`);
+      else 
+        console.info(`[INFO] Code file ${localCodePath} succesfully prepared.`)
+    });
+  }
+
+  // input data file
   let localDataPath = `${__dirname}\\..\\data\\${settings.Settings.YEAR}\\${name}`;
-  if (!fs.existsSync(localDataPath)){
-      fs.mkdirSync(localDataPath);
+  if (!fs.existsSync(localDataPath)) {
+    fs.mkdirSync(localDataPath, { "recursive": true });
   }
   localDataPath += `\\input.txt`;
 
-  let localCodePath = `${__dirname}\\..\\src\\${settings.Settings.YEAR}\\${name}`;
-  if (!fs.existsSync(localCodePath)){
-      fs.mkdirSync(localCodePath);
-  }
-  localCodePath += `\\code.ts`;
-  let code = fs.readFileSync(`${__dirname}\\..\\src\\code-template.ts`).toString();
-  code = code.replace(/day00/g, name);
-  fs.writeFile(localCodePath, code, (err) => {
-    if (err)
-      console.log(`Error: ${err}`);
-  });
-
-
-  console.log(`Attempting to download from ${options.path} to ${localDataPath}`);
+  console.info(`[INFO] Attempting to download from ${options.path} to ${localDataPath}`);
   https.get(options, (res) => {
-      const fileStream = fs.createWriteStream(localDataPath);
-      res.pipe(fileStream);
-      fileStream.on('finish', () => {
-          fileStream.close();
-          console.log('Download complete!');
-          process.exit();
-      });
+    const fileStream = fs.createWriteStream(localDataPath);
+    res.pipe(fileStream);
+    fileStream.on('finish', () => {
+      fileStream.close();
+      console.info('[INFO] Download complete!');
+      process.exit();
+    });
   });
 };
